requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiohttp==3.10.5
APScheduler==3.11.0
discord.py==2.5.2
pillow==10.4.0
pixelsort==1.0.1
pymep==1.0.7
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2024.1
requests==2.32.3
scipy==1.13.1
wavelink==3.4.1
